decimal multiplication

+++ KnowledgeCheck1_Calculator/Program.cs
@@ -55,7 +55,7 @@ namespace KnowledgeCheck1_Calculator
                     var multiplyNumber1 = Console.ReadLine();
                     var multiplyNumber2 = Console.ReadLine();
 
-                    if (int.TryParse(multiplyNumber1, out int mulNumOne) && int.TryParse(multiplyNumber2, out int mulNumTwo))
+                    if (double.TryParse(multiplyNumber1, out double mulNumOne) && double.TryParse(multiplyNumber2, out double mulNumTwo))
                     {
                         Console.Write($"{multiplyNumber1} * {multiplyNumber2} = ");
                         Console.Write(calculator.Multiply(mulNumOne, mulNumTwo));
